18/2

+++ functionality/_log_out.php
@@ -1,19 +1,23 @@
 <?php
 session_start();
-if(isset($_GET['get_pass']) && $_GET['get_pass'] === "khulJaSimSim") {
+if(isset($_GET['key_pass']) && $_GET['key_pass'] === "khulJaSimSim") {
     session_destroy();
     session_abort();
 }
 ?>
 <script>
     var URL_Param = new URLSearchParams(window.location.search);
-    console.log(URL_Param);
-    var requst= indexedDB.deleteDatabase('Botsapp');
+    if(URL_Param.get('key_pass') && (URL_Param.get('key_pass') == "khulJaSimSim")){
+        var requst= indexedDB.deleteDatabase('Botsapp');
 
-    requst.onsuccess = (event) => {
+        requst.onsuccess = (event) => {
+            window.location.assign('/');
+        }
+        requst.onerror = (event) => {
+            console.log(event);
+        }
+    }else{
         window.location.assign('/');
     }
-    requst.onerror = (event) => {
-        console.log(event);
-    }
+    
 </script>
